+++ packages/core/constants.ts
@@ -1,6 +1,5 @@
-export const GA_CODE = process.env.GA_CODE;
 export const VERSION = process.env.VERSION;
 export const EXT_NAME = process.env.EXT_NAME;
 export const AUTHOR = process.env.AUTHOR;
-export const BUGSNAG_API_KEY = process.env.BUGSNAG_API_KEY;
 export const ENV = process.env.ENV;
+export const AI_KEY = 'd739ccec-c4e8-45e0-9d14-3b43d57a1db7';