+++ packages/core/utils/index.ts
@@ -1,5 +1,6 @@
 import * as query from './query';
 import * as persistence from './persistence';
+import { ConnectionCredentials, SerializedConnection } from '../interface';
 
 export * from './get-home';
 export * from './replacer';
@@ -7,5 +8,10 @@ export * from './telemetry';
 export * from './timer';
 
 export function sortText(a: string, b: string) { return a.toString().localeCompare(b.toString()); }
+
+export function getDbId(c: SerializedConnection | ConnectionCredentials): string | null {
+  if (!c) return null;
+  return `${c.name}#${c.database}#${c.dialect}`;
+}
 
 export { query, persistence };